commit before pull check
