+++ src/models/benefit.js
@@ -14,7 +14,7 @@ class BenefitModel {
 
   static create = (userBenefitData) => {
     return fetch(`${REACT_APP_API_URL}/allbenefits`, {
-      method: "POST",
+      method: "GET",
       headers: {
         "Content-Type": "application/json"
       },
